requirements.txt
# Core Dependencies - Updated for OpenAI
openai>=1.12.0
# Removed: anthropic, langchain-anthropic

# Web Scraping & Automation
selenium>=4.16.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0

# NLP & ML
sentence-transformers>=2.3.0
scikit-learn>=1.3.0
spacy>=3.7.0

# Database
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.0
tqdm>=4.66.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0
aiohttp>=3.9.0

# Logging & Monitoring
loguru>=0.7.0

# Additional for scheduling
schedule>=1.2.0